refactor: extract the merge key-rename planner from the delete behaviors
The backspace and forward-delete merge behaviors planned their
colliding-key renames inside `behavior.abstract.delete.ts`, coupled to
behavior actions (`raise`). The planning itself — which keys collide,
what they become, which `marks` follow a renamed markDef — is
layer-neutral, and the range-delete merge at the operation layer needs
the same plan.

Move the pure planner to `internal-utils/plan-merge-key-renames.ts`,
returning the renamed block plus neutral child and markDef rename
lists; the behavior builds its `raise` actions from those lists.

Net behavior unchanged: colliding and non-colliding merges emit
byte-identical patches, pinned by the existing
`block-merge-duplicate-keys` suite and wire-catalogue fixtures, both
untouched.
